# web/src/api/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "GameEvent.Web | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/seasons/{seasonId}/feed": {
      "get": {
        "tags": [
          "Seasons"
        ],
        "parameters": [
          {
            "name": "seasonId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "before",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedView"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/{userId}": {
      "get": {
        "tags": [
          "Accounts"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileView"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/pool/{gameId}/runs": {
      "get": {
        "tags": [
          "Pool"
        ],
        "parameters": [
          {
            "name": "gameId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GameRunView"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Difficulty": {
        "enum": [
          "easy",
          "normal",
          "hard",
          "extreme"
        ]
      },
...
      "FeedEntryView": {
        "required": [
          "sequence",
          "commandId",
          "occurredAt",
          "type",
          "data",
          "author",
          "undone"
        ],
        "type": "object",
        "properties": {
          "sequence": {
            "type": "integer",
            "format": "int64"
          },
          "commandId": {
            "type": "string",
            "format": "uuid"
          },
          "occurredAt": {
            "type": "string",
            "format": "date-time"
          },
          "type": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/JsonElement"
          },
          "author": {
            "type": [
              "null",
              "string"
            ]
          },
          "undone": {
            "type": "boolean"
          }
        },
        "description": "One event of the season's feed (GLOSSARY «Лента», D-124): its place in the log, the command it came from, when, the\nevent type and its data as the current format reads it (older versions upcast), who sent the command (none for the\nscheduler) and whether the admin undid it. The screen turns a type and its data into a line of text."
      },
      "FeedGameView": {
        "required": [
          "id",
          "title",
          "hasPage"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "hasPage": {
            "type": "boolean"
          },
          "cover": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/FileLinkView"
              }
            ]
          }
        },
        "description": "A game an event of the page names; `hasPage` — the viewer may open its page (a deleted game is the admin's);\n`cover` — the pool's cover, as the season screen shows it (D-222)."
      },
      "FeedPlayerView": {
        "required": [
          "id",
          "userId",
          "name",
          "avatar",
          "hasProfile",
          "token"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "avatar": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/FileLinkView"
              }
            ]
          },
          "hasProfile": {
            "type": "boolean"
          },
          "token": {
            "type": "integer",
            "format": "int32"
          }
        },
        "description": "A player of the season, for the feed's lines (H5, D-150): the name, the account for the profile link (`hasProfile` —\nfalse once the account is deleted), the avatar and `token` — the colour of their token in the season (PlayerTokens)."
      },
      "FeedRunView": {
        "required": [
          "id",
          "gameId"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "gameId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "description": "A run an event of the page names, with its game: most events of a run carry only its id."
      },
      "FeedView": {
        "required": [
          "entries",
          "nextBefore",
          "players",
          "games",
          "runs"
        ],
        "type": "object",
        "properties": {
          "entries": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FeedEntryView"
            }
          },
          "nextBefore": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int64"
          },
          "players": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FeedPlayerView"
            }
          },
          "games": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FeedGameView"
            }
          },
          "runs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FeedRunView"
            }
          }
        },
        "description": "A page of the feed, newest first; `nextBefore` — the sequence to ask before for the next page, none at the start of\nthe log. `players` — every player of the season; `games` and `runs` — those the page's events name (D-150)."
      },
...
      "FileLinkView": {
        "required": [
          "id",
          "url",
          "thumbnailUrl"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "url": {
            "type": "string"
          },
          "thumbnailUrl": {
            "type": "string"
          }
        },
        "description": "A stored file where a proof or a page shows it (D-116): the picture and its thumbnail."
      },
...
      "GameRunView": {
        "required": [
          "runId",
          "seasonId",
          "seasonName",
          "playerId",
          "userId",
          "playerName",
          "status",
          "difficulty",
          "hours",
          "completedAt",
          "rating",
          "reviewText",
          "token"
        ],
        "type": "object",
        "properties": {
          "runId": {
            "type": "string",
            "format": "uuid"
          },
          "seasonId": {
            "type": "string",
            "format": "uuid"
          },
          "seasonName": {
            "type": "string"
          },
          "playerId": {
            "type": "string",
            "format": "uuid"
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "playerName": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/RunStatus"
          },
          "difficulty": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/Difficulty"
              }
            ]
          },
          "hours": {
            "type": [
              "null",
              "number"
            ],
            "format": "double"
          },
          "completedAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "rating": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "reviewText": {
            "type": [
              "null",
              "string"
            ]
          },
          "token": {
            "type": "integer",
            "format": "int32"
          }
        },
        "description": "A run of a game in any season, with its review if there is one (the game page, D-124). `token` — the player's place\nin that season's list of players, their token colour there (D-150)."
      },
...
      "HttpValidationProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        }
      },
...
      "JsonElement": { },
...
      "ProfileReviewView": {
        "required": [
          "runId",
          "gameId",
          "gameTitle",
          "seasonId",
          "seasonName",
          "rating",
          "text",
          "completedAt"
        ],
        "type": "object",
        "properties": {
          "runId": {
            "type": "string",
            "format": "uuid"
          },
          "gameId": {
            "type": "string",
            "format": "uuid"
          },
          "gameTitle": {
            "type": "string"
          },
          "seasonId": {
            "type": "string",
            "format": "uuid"
          },
          "seasonName": {
            "type": "string"
          },
          "rating": {
            "type": "integer",
            "format": "int32"
          },
          "text": {
            "type": [
              "null",
              "string"
            ]
          },
          "completedAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          }
        },
        "description": "A review the user left: the game, the season, the rating and the text (SPEC «Отзыв»)."
      },
      "ProfileSeasonView": {
        "required": [
          "seasonId",
          "seasonName",
          "status",
          "playerId",
          "points",
          "place",
          "token"
        ],
        "type": "object",
        "properties": {
          "seasonId": {
            "type": "string",
            "format": "uuid"
          },
          "seasonName": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/SeasonStatus"
          },
          "playerId": {
            "type": "string",
            "format": "uuid"
          },
          "points": {
            "type": "integer",
            "format": "int32"
          },
          "place": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "token": {
            "type": "integer",
            "format": "int32"
          }
        },
        "description": "A season a user took part in, with their points and, once the season finished, their place. `token` — their place in\nthe season's list of players (by name), which the screens colour their token by (D-150)."
      },
      "ProfileView": {
        "required": [
          "id",
          "name",
          "avatar",
          "seasons",
          "reviews",
          "completed"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "avatar": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/FileLinkView"
              }
            ]
          },
          "seasons": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProfileSeasonView"
            }
          },
          "reviews": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProfileReviewView"
            }
          },
          "completed": {
            "type": "integer",
            "format": "int32"
          }
        },
        "description": "A user's profile (D-124): the name, the avatar, the seasons they played and their reviews. `completed` — games\ncompleted and not rejected, over all seasons."
      },
...
      "RunStatus": {
        "enum": [
          "playing",
          "completed",
          "dropped",
          "techRerolled",
          "rejected"
        ]
      },
...
      "SeasonStatus": {
        "enum": [
          "draft",
          "active",
          "closing",
          "finished",
          "archived"
        ],
        "description": "Season lifecycle (GLOSSARY «Статус сезона»): draft → active → closing → finished → archived."
      },
...
    }
  },
...
}